Fix a crash when deleting one message only for me
Fix a bug due to a fork change where the progress dialog would not pop up when only deleting one message. That change caused single message deletion for me to occur in the main/ui thread. When deleting an image or video message, a function requires being called from a background thread, and this will end in the app crashing.